implement in-combat show/hide & avoid taint
this is essentially a reimplementation that just uses show/hide on the
relevant frames, and avoids touching multiactionbar configuration.
modifying the config through the interface options frames is not doable
in the restricted environment since Click()ing the checkboxes is not
available, and we can't touch the globals that make up those settings
either.